+++ CHANGELOG.md
@@ -14,6 +14,7 @@
 - Added a check for LifeForce-Empath having an active life-force gene (#143), which likely threw errors being active despite children not having life-force yet.
 - Added a map-check for LifeForce Empaths. Won't work off-map. (#120)
 - Added another Patch for `Blocked Masturbation` that disables the ThinkNode like Sexperience does (#127,#147). Thanks @TinyMechanoid333 !
+- Fixed males spawning with small-female breasts (#138) by adding a detailed check `should have breasts`
 
 # 2.2.0 (23-07-2024)